Add ShortestPathApiTest
